soi_bench_agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "soi-bench-agent"
version = "0.1.0"
description = "A multi-turn benchmark-grounded recommendation agent for spatial omics integration methods."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "setuptools>=82.0.1",
]

[project.scripts]
soi-bench-agent = "soi_bench_agent.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
